rm duplicate graph
