+++ packages/app/src/library/Graphs/types.ts
@@ -3,11 +3,7 @@
 
 import type BigNumber from 'bignumber.js'
 import type { AnyApi } from 'common-types'
-import type {
-  NominatorReward,
-  PoolReward,
-  ValidatorEraPoints,
-} from 'plugin-staking-api/types'
+import type { NominatorReward, PoolReward } from 'plugin-staking-api/types'
 
 export interface BondedProps {
   active: BigNumber
@@ -15,11 +11,6 @@ export interface BondedProps {
   unlocking: BigNumber
   unlocked: BigNumber
   inactive: boolean
-}
-
-export interface EraPointsProps {
-  items: ValidatorEraPoints[]
-  height: number
 }
 
 export interface PayoutBarProps {
